requirements.txt
httpx>=0.27,<0.28
pandas>=2.2,<3.0
python-dotenv>=1.0,<2.0
schedule>=1.2,<2.0
streamlit>=1.40,<2.0
altair>=5.0,<6.0
pytest>=8.0,<9.0
